Report-Generator-IP-main/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
pandas==2.1.4
python-docx==1.1.2
docxtpl==0.19.1
matplotlib==3.7.2
numpy==1.24.3
python-dateutil==2.8.2
docxcompose==1.4.0
Pillow==10.0.1
starlette==0.46.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
requests==2.31.0
jinja2==3.1.2
openpyxl==3.1.2
babel==2.12.1
pydantic==2.4.2
gunicorn==21.2.0 
